Change many unsigned to signed

+++ AeroMessages/GSS/V66/Character/Command/AcquireWeaponTarget.cs
@@ -8,7 +8,7 @@ namespace AeroMessages.GSS.V66.Character.Command
     {
         public uint Unk1;
         public ulong Unk2;
-        public uint Unk3;
-        public byte Unk4;
+        public float Unk3;
+        public sbyte Unk4;
     }
 }
